feat: Adding `failOnMotionCorrelationFailedAtTheEnd` option to control failure on motion correlation (default = true)
When this option is `false`, a failure in the final motion correlation check will no longer cause the entire liveness verification to fail. The result of the check will still be available in the metadata.

Additionally, the session will now reset if the face is lost during any non-movement challenges.