Refactor: Improve motion correlation logic
The method for calculating device motion in `verifyMotionCorrelation` has been updated. Previously, it used the range of accelerometer vector magnitudes.

This has been changed to use the standard deviation for each accelerometer axis (X, Y, Z) independently. The maximum of these standard deviations is now used to represent device motion. This approach is more robust as it correctly detects rotational movements that might not significantly alter the vector magnitude.